+++ src/num/sign.rs
@@ -14,6 +14,7 @@ use crate::{
 };
 
 /// Represents the sign of a number.
+#[must_use]
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Hash)]
 pub enum Sign {
     /// A negative sign (-).
@@ -41,7 +42,6 @@ macro_rules! impl_into_sign {
     (@int: $int:ty) => {
         impl From<$int> for Sign {
             /// Returns `None` if 0, `Positive` if > 0 and `Negative` if < 0.
-            #[must_use]
             fn from(n: $int) -> Sign {
                 match n {
                     0 => Sign::None,
@@ -57,7 +57,6 @@ macro_rules! impl_into_sign {
     (@float: $float:ty) => {
         impl From<$float> for Sign {
             /// Returns `None` if 0.0, `Positive` if > 0 and `Negative` if < 0.
-            #[must_use]
             fn from(n: $float) -> Sign {
                 if n.is_sign_positive() {
                     Sign::Positive
@@ -71,7 +70,6 @@ macro_rules! impl_into_sign {
     (bool) => {
         impl From<bool> for Sign {
             /// Returns `Positive` if `true` and `Negative` if `false`.
-            #[must_use]
             fn from(n: bool) -> Sign {
                 match n {
                     true => Sign::Positive,
@@ -128,7 +126,6 @@ macro_rules! impl_from_sign {
     (@float: $float:ty) => {
         impl From<Sign> for $float {
             /// Returns 0.0 if `None`, 1.0 if `Positive` and -1.0 if `Negative`.
-            #[must_use]
             fn from(s: Sign) -> $float {
                 match s {
                     Sign::None => 0.0,